# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SelfMBA Marketing Portfolio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SurveyInputHowFound = typeof SurveyInputHowFound[keyof typeof SurveyInputHowFound];


export const SurveyInputHowFound = {
  linkedin: 'linkedin',
  google: 'google',
  referral: 'referral',
  social_media: 'social_media',
  direct: 'direct',
  other: 'other',
} as const;

export type SurveyInputMostValuableSection = typeof SurveyInputMostValuableSection[keyof typeof SurveyInputMostValuableSection];


export const SurveyInputMostValuableSection = {
  brand_identity: 'brand_identity',
  target_audience: 'target_audience',
  positioning: 'positioning',
  marketing_mix: 'marketing_mix',
  consumer_insights: 'consumer_insights',
  campaign_strategy: 'campaign_strategy',
  overall: 'overall',
} as const;

export type SurveyInputMarketingBackground = typeof SurveyInputMarketingBackground[keyof typeof SurveyInputMarketingBackground];


export const SurveyInputMarketingBackground = {
  student: 'student',
  professional: 'professional',
  entrepreneur: 'entrepreneur',
  recruiter: 'recruiter',
  curious: 'curious',
  other: 'other',
} as const;

export interface SurveyInput {
  /** @nullable */
  respondentName?: string | null;
  /** @nullable */
  respondentEmail?: string | null;
  howFound?: SurveyInputHowFound;
  /**
     * @minimum 1
     * @maximum 5
     */
  overallRating: number;
  mostValuableSection: SurveyInputMostValuableSection;
  wouldRecommend: boolean;
  wouldHire?: boolean;
  /** @nullable */
  openFeedback?: string | null;
  marketingBackground?: SurveyInputMarketingBackground;
}

export interface SurveyResponse {
  id: number;
  /** @nullable */
  respondentName?: string | null;
  /** @nullable */
  respondentEmail?: string | null;
  /** @nullable */
  howFound?: string | null;
  overallRating: number;
  mostValuableSection: string;
  wouldRecommend: boolean;
  /** @nullable */
  wouldHire?: boolean | null;
  /** @nullable */
  openFeedback?: string | null;
  /** @nullable */
  marketingBackground?: string | null;
  submittedAt: string;
}

export interface DistributionItem {
  label: string;
  count: number;
  percentage: number;
}

export interface FeedbackSnippet {
  id: number;
  text: string;
  rating: number;
  /** @nullable */
  name?: string | null;
  submittedAt: string;
}

export interface SurveyAnalytics {
  totalResponses: number;
  averageRating: number;
  recommendRate: number;
  hireRate: number;
  ratingDistribution: DistributionItem[];
  sectionPopularity: DistributionItem[];
  howFoundDistribution: DistributionItem[];
  backgroundDistribution: DistributionItem[];
  recentFeedback: FeedbackSnippet[];
}