# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SelfMBA Marketing Portfolio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /survey/responses:
    post:
      operationId: submitSurvey
      tags: [survey]
      summary: Submit a survey response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SurveyInput"
      responses:
        "201":
          description: Survey submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurveyResponse"
        "400":
          description: Validation error
    get:
      operationId: listSurveyResponses
      tags: [survey]
      summary: List all survey responses
      responses:
        "200":
          description: List of responses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SurveyResponse"
...
components:
  schemas:
...
    SurveyInput:
      type: object
      required:
        - overallRating
        - mostValuableSection
        - wouldRecommend
      properties:
        respondentName:
          type: ["string", "null"]
        respondentEmail:
          type: ["string", "null"]
        howFound:
          type: string
          enum: [linkedin, google, referral, social_media, direct, other]
        overallRating:
          type: integer
          minimum: 1
          maximum: 5
        mostValuableSection:
          type: string
          enum: [brand_identity, target_audience, positioning, marketing_mix, consumer_insights, campaign_strategy, overall]
        wouldRecommend:
          type: boolean
        wouldHire:
          type: boolean
        openFeedback:
          type: ["string", "null"]
        marketingBackground:
          type: string
          enum: [student, professional, entrepreneur, recruiter, curious, other]
    SurveyResponse:
      type: object
      required:
        - id
        - overallRating
        - mostValuableSection
        - wouldRecommend
        - submittedAt
      properties:
        id:
          type: integer
        respondentName:
          type: ["string", "null"]
        respondentEmail:
          type: ["string", "null"]
        howFound:
          type: ["string", "null"]
        overallRating:
          type: integer
        mostValuableSection:
          type: string
        wouldRecommend:
          type: boolean
        wouldHire:
          type: ["boolean", "null"]
        openFeedback:
          type: ["string", "null"]
        marketingBackground:
          type: ["string", "null"]
        submittedAt:
          type: string
...