# lib/api-zod/src/generated/types/surveyInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SelfMBA Marketing Portfolio API
 * OpenAPI spec version: 0.1.0
 */
import type { SurveyInputHowFound } from './surveyInputHowFound';
import type { SurveyInputMarketingBackground } from './surveyInputMarketingBackground';
import type { SurveyInputMostValuableSection } from './surveyInputMostValuableSection';

export interface SurveyInput {
  /** @nullable */
  respondentName?: string | null;
  /** @nullable */
  respondentEmail?: string | null;
  howFound?: SurveyInputHowFound;
  /**
     * @minimum 1
     * @maximum 5
     */
  overallRating: number;
  mostValuableSection: SurveyInputMostValuableSection;
  wouldRecommend: boolean;
  wouldHire?: boolean;
  /** @nullable */
  openFeedback?: string | null;
  marketingBackground?: SurveyInputMarketingBackground;
}