# lib/api-zod/src/generated/types/surveyResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SelfMBA Marketing Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface SurveyResponse {
  id: number;
  /** @nullable */
  respondentName?: string | null;
  /** @nullable */
  respondentEmail?: string | null;
  /** @nullable */
  howFound?: string | null;
  overallRating: number;
  mostValuableSection: string;
  wouldRecommend: boolean;
  /** @nullable */
  wouldHire?: boolean | null;
  /** @nullable */
  openFeedback?: string | null;
  /** @nullable */
  marketingBackground?: string | null;
  submittedAt: string;
}